Split tlint-wasm to a different crate
