+++ Cargo.toml
@@ -4,25 +4,17 @@ version = "0.9.4"
 edition = "2021"
 
 [lib]
-crate-type = ["cdylib", "rlib"]
-
-[profile.release]
-# Optimize wasm size
-lto = true          # turn on Link-Time Optimizations
-codegen-units = 1   # trade compile time with maximum optimization
-opt-level = "s"     # optimize for size
+name = "tlint"
 
 [dependencies]
 clap = { version = "4.5.19", features = ["derive"] }
 colored = "2"
 jsonschema = { version= "0.19.1", default-features = false }
 yaml-rust = "0.4"
-rhai = { version = "1.13.0", features = ["only_i32", "internals", "wasm-bindgen"] }
+rhai = { version = "1.13.0", features = ["internals"] }
 serde = {version = "1.0.147", features = ["derive"] }
 serde_yaml = "0.9.14"
 serde_json = "1.0.87"
-wasm-bindgen = { version = "0.2.88", features = ["serde-serialize"] }
-serde-wasm-bindgen = "0.4"
 
 [dev-dependencies]
 assert_cmd = "2.0.12"
