+++ www/.gitignore
@@ -1,2 +1,4 @@
-node_modules
-dist
+/node_modules
+/dist
+/pkg
+/target
