+++ www/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "tlint-wasm"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[profile.release]
+# Optimize wasm size
+lto = true          # turn on Link-Time Optimizations
+codegen-units = 1   # trade compile time with maximum optimization
+opt-level = "s"     # optimize for size
+
+[dependencies]
+tlint = { path = "../" }
+rhai = { version = "1.13.0", features = ["only_i32", "internals", "wasm-bindgen"] }
+serde = {version = "1.0.147", features = ["derive"] }
+serde_yaml = "0.9.14"
+serde_json = "1.0.87"
+wasm-bindgen = { version = "0.2.88", features = ["serde-serialize"] }
+serde-wasm-bindgen = "0.4"
