+++ www/package.json
@@ -12,7 +12,7 @@
     "@codemirror/lang-yaml": "^6.1.1",
     "codemirror": "^6.0.1",
     "comlink": "^4.4.2",
-    "tlint": "file:../pkg"
+    "tlint": "file:pkg"
   },
   "devDependencies": {
     "copy-webpack-plugin": "^5.1.2",
