+++ www/webpack.config.js
@@ -5,7 +5,7 @@ const dist = path.resolve(__dirname, "dist");
 module.exports = {
   // mode: "production",
   mode: "development",
-  entry: "./src/bootstrap.js",
+  entry: "./js/bootstrap.js",
   output: {
     path: dist,
     filename: "main.js",
